Updated "Pair with Target Sum" (Two Pointers)
the original "two sum" which is linked in the readme is not an ideal example for two sum, as here the input array is not sorted. hash maps will be a more appropriate pattern for this question. a better example would be "two sum - II", where the input array is sorted and two pointers approach can be directly applied.

even the original grokking the coding interview has sorted array for the problem.